+++ test/hexapdf/test_importer.rb
@@ -128,6 +128,16 @@ describe HexaPDF::Importer do
       refute_same(dst_obj.data.stream, src_obj.data.stream)
     end
 
+    it "duplicates the stream if it is a FiberDoubleForString, e.g. when using Canvas" do
+      src_page = @source.pages[0]
+      src_page.canvas.line_width(10)
+      dst_page = @importer.import(src_page)
+      refute_same(dst_page, src_page)
+      refute_same(dst_page[:Contents].data.stream, src_page[:Contents].data.stream)
+      src_page.canvas.line_width(20)
+      assert_equal("10 w\n", dst_page.contents)
+    end
+
     it "does not import objects of type Catalog or Pages" do
       @obj[:catalog] = @source.catalog
       @obj[:pages] = @source.catalog.pages